+++ CHANGELOG.md
@@ -1,3 +1,5 @@
-11May23 13:00 - Updated to automatically load all resources into Assets folder upon right-click-> add sxr_prefab
+1.0.1 - 11May23 13:00 - Updated to automatically load all resources into Assets folder upon right-click-> add sxr_prefab
 
-14May23 18:19 - Update ModifyShader and disable duplicating sXR shaders. Add physics modifier
+1.0.2 - 14May23 18:19 - Update ModifyShader and disable duplicating sXR shaders. Add physics modifier
+
+1.0.3 - 16May23 01:29 - Multiple Updates to revert ShaderHandler back to using Graphics.Blit(), enables stereoscopic rendering with multi-pass. Added rotate object functions